requirements.txt
scrapy ~= 2.11.1
gspread ~= 6.1.0
google-api-python-client 